% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create a codebook for the survey}
\usage{
codebook(all = FALSE)
}
\arguments{
\item{all}{tabulate all the variables?}
}
\value{
A list of tables.
}
\description{
Create a codebook for the survey
}
\examples{
set_survey(namcs2019sv)
codebook()
}
