% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{set_mode}
\alias{set_count_1k}
\alias{set_count_int}
\alias{set_output}
\title{Deprecated functions}
\usage{
set_mode(mode = "general")

set_count_1k()

set_count_int()

set_output(drop_na = NULL, max_levels = NULL, csv = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link[=set_opts]{set_opts()}} instead of any of the following: \code{set_mode()}, \code{set_count_1k()},
\code{set_count_int()}, \code{set_output()}.
}
\keyword{internal}
