% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sas_namcs.R
\name{import_sas_namcs2019puf}
\alias{import_sas_namcs2019puf}
\title{Import SAS data into R, with survey design variables the same as those in NAMCS 2019 PUF.}
\usage{
import_sas_namcs2019puf(sas_data, sas_formats_data, r_out)
}
\arguments{
\item{sas_data}{SAS survey data file}

\item{sas_formats_data}{SAS formats data file (produced with \code{CNTLOUT} option of \verb{PROC FORMAT})}

\item{r_out}{name of a new R data file}
}
\value{
(Nothing.)
}
\description{
Import SAS data into R, with survey design variables the same as those in NAMCS 2019 PUF.
}
\examples{
\dontrun{
import_sas_namcs2019puf(
sas_data = "namcs2019_sas.sas7bdat"
, sas_formats_data = "formats_dataset.sas7bdat"
, r_out = "namcs_2019_puf.rds")
}
}
\seealso{
Other import: 
\code{\link{import_sas}()}
}
\concept{import}
