% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{namcs2019}
\alias{namcs2019}
\title{National Ambulatory Medical Care Survey (NAMCS) 2019 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 8250 rows and 1095 columns.
}
\usage{
namcs2019
}
\description{
Visits to physicians. Visits are not the same as individuals, as one
individual can make multiple visits.
}
\details{
A survey design object with 1,095 variables and 8,250 observations.
}
\keyword{datasets}
