% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{namcs2019sv}
\alias{namcs2019sv}
\alias{namcs2019sv_df}
\title{Selected variables from the National Ambulatory Medical Care Survey (NAMCS) 2019 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 8250 rows and 33 columns.

An object of class \code{data.frame} with 8250 rows and 33 columns.
}
\usage{
namcs2019sv

namcs2019sv_df
}
\description{
Selected variables from a data system of visits to office-based physicians.
Note that the unit of observation is visits, not patients - this distinction
is important since a single patient can make multiple visits.
}
\details{
\code{namcs2019sv_df} is a data frame.

\code{namcs2019sv} is a survey object created from \code{namcs2019sv_df}
using \code{survey::svydesign()}.
}
\keyword{datasets}
