% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{nhis2024a}
\alias{nhis2024a}
\title{National Health Interview Survey (NHIS) 2024 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 32577 rows and 16 columns.
}
\source{
\itemize{
\item Data ("Sample adult interview"): \url{https://www.cdc.gov/nchs/nhis/documentation/2024-nhis.html}
}
}
\usage{
nhis2024a
}
\description{
NHIS is a national survey that monitors the health of the U.S. population. This
survey object contains selected variables and is limited to observations on
individuals aged 18+ only.
}
\keyword{datasets}
