% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{opts}
\alias{opts}
\title{\code{opts}: options}
\format{
An object of class \code{list} of length 3.
}
\usage{
opts
}
\description{
\code{opts}: options
}
\examples{
opts$out$fname = "output.html"
}
\keyword{datasets}
