% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettysurvey-package.R
\docType{package}
\name{prettysurvey-options}
\alias{prettysurvey-options}
\title{Package options}
\description{
Run \code{\link[=show_options]{show_options()}} to see available options.
}
\seealso{
Other options: 
\code{\link{set_count_1k}()},
\code{\link{set_survey}()},
\code{\link{show_options}()}
}
\concept{options}
