% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettysurvey-package.R
\docType{package}
\name{prettysurvey-package}
\alias{prettysurvey}
\alias{prettysurvey-package}
\title{prettysurvey: Streamlined output from the `survey` package}
\description{
Streamlined output from the `survey` package.
}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov}

}
\keyword{internal}
