% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{rccsu2018}
\alias{rccsu2018}
\title{National Study of Long-Term Care Providers (NSLTCP) Residential Care Community (RCC) Services User (SU) 2018 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 904 rows and 81 columns.
}
\source{
\itemize{
\item SAS data: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Datasets/NPALS/final2018rcc_su_puf.sas7bdat}
\item Documentation: \url{https://www.cdc.gov/nchs/npals/RCCresident-readme03152021vr.pdf}
\item Codebook: \url{https://www.cdc.gov/nchs/data/npals/final2018rcc_su_puf_codebook.pdf}
}
}
\usage{
rccsu2018
}
\description{
A data system of RCC residents.
}
\keyword{datasets}
