% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_excel.R
\name{save_excel}
\alias{save_excel}
\title{Save tables and charts to an Excel workbook}
\usage{
save_excel(file = NULL, overwrite = FALSE, .temp = FALSE)
}
\arguments{
\item{file}{file name, or \code{NULL} (default) to generate one automatically.}

\item{overwrite}{overwrite file if it exists?}

\item{.temp}{save to a temporary location?}
}
\value{
(Nothing.)
}
\description{
First, initiate Excel printing with \code{set_opts( output = "Excel" )}. Then, use
the various tabulation functions to "print" / generate the tables and charts.
Finally, call \code{save_excel()} to save all of these tables and charts to a
single Excel file, which consists of one or more spreadsheets, one for each table or chart.
}
\details{
By default, a file name is generated automatically.

Before using Excel printing, be sure to install these packages: \code{openxlsx2} and \code{mschart}.
}
\examples{
set_survey(namcs2019sv)
set_opts(output = "excel")
total()
tab("AGER")
save_excel(file = "my_workbook.xlsx", .temp = TRUE)
set_opts(output = "auto")
}
\seealso{
Other print: 
\code{\link{print.surveytable_table}()}
}
\concept{print}
