% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_output.R
\name{set_output}
\alias{set_output}
\alias{show_output}
\title{Set output defaults}
\usage{
set_output(drop_na = NULL, max_levels = NULL, csv = NULL)

show_output()
}
\arguments{
\item{drop_na}{drop missing values (\code{NA})? Categorical variables only.}

\item{max_levels}{a categorical variable can have at most this many levels. Used to avoid printing huge tables.}

\item{csv}{name of a CSV file or "" to turn off CSV output}
}
\value{
(Nothing.)
}
\description{
\code{show_output()} shows the current defaults.
}
\examples{
tmp_file = tempfile(fileext = ".csv")
suppressMessages( set_output(csv = tmp_file) )
tab("AGER")
set_output(csv = "") # Turn off CSV output
}
\seealso{
Other options: 
\code{\link{set_count_1k}()},
\code{\link{set_survey}()},
\code{\link{show_options}()},
\code{\link{surveytable-options}}
}
\concept{options}
