% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_subset.R
\name{survey_subset}
\alias{survey_subset}
\title{Subset a survey, while preserving variable labels}
\usage{
survey_subset(design, subset, label)
}
\arguments{
\item{design}{a survey object}

\item{subset}{an expression specifying the sub-population}

\item{label}{survey label of the newly created survey object}
}
\value{
a new survey object
}
\description{
Subset a survey, while preserving variable labels
}
\examples{
children = survey_subset(namcs2019sv, AGE < 18, "Children < 18")
set_survey(children)
tab("AGER")
}
