% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytable.R
\docType{package}
\name{surveytable-package}
\alias{surveytable}
\alias{surveytable-package}
\title{surveytable: Formatted Survey Estimates}
\description{
Short and understandable commands that generate tabulated, formatted, and rounded survey estimates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/surveytable/}
  \item \url{https://github.com/CDCgov/surveytable}
}

}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov} (\href{https://orcid.org/0000-0002-6408-7745}{ORCID})

}
\keyword{internal}
