% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_by.R
\name{tab_subset}
\alias{tab_subset}
\alias{tab_cross}
\alias{var_cross}
\title{Tabulate subsets or interactions}
\usage{
tab_subset(
  design,
  vr,
  vrby,
  lvls = c(),
  max.levels = getOption("prettysurvey.out.max_levels"),
  screen = getOption("prettysurvey.out.screen"),
  out = getOption("prettysurvey.out.fname")
)

tab_cross(
  design,
  vr,
  vrby,
  max.levels = getOption("prettysurvey.out.max_levels"),
  screen = getOption("prettysurvey.out.screen"),
  out = getOption("prettysurvey.out.fname")
)

var_cross(design, newvr, vr, vrby)
}
\arguments{
\item{design}{survey design}

\item{vr}{variable to tabulate}

\item{vrby}{use this variable to subset the survey}

\item{lvls}{(optional) only show these levels of \code{vrby}}

\item{max.levels}{a categorical variable can have at most this many levels. Used to avoid printing huge tables.}

\item{screen}{print to the screen?}

\item{out}{file name of CSV file}

\item{newvr}{name of the new variable to be created}
}
\value{
\itemize{
\item \code{tab_subset}: a list of \code{data.frame} tables.
\item \code{tab_cross}: a \code{data.frame} table.
\item \code{var_cross}: an updated \code{survey.design}.
}
}
\description{
Create subsets of the survey using one variable, and tabulate another
variable within each of the subsets. Interact two variables and tabulate.
}
\details{
\code{tab_subset} creates subsets using the levels of \code{vrby}, and tabulates
\code{vr} in each subset. Optionally, only use the \code{lvls} levels of \code{vrby}.

\code{tab_cross} crosses or interacts \code{vr} and \code{vrby} and tabulates the new
variable. Tables created using \code{tab_subset} and \code{tab_cross} have the same
counts but different percentages. With \code{tab_subset}, percentages within each
subset add up to 100\%. With \code{tab_cross}, percentages across the entire survey
add up to 100\%.

\code{var_cross} creates the new variable, returns the updates survey,
but does not tabulate the new variable. Use \code{\link[=tab]{tab()}} to tabulate it.
}
\examples{
# For each SEX, tabulate AGER
tab_subset(namcs2019, "AGER", "SEX")

# Same counts as tab_subset(), but different percentages.
tab_cross(namcs2019, "AGER", "SEX")

# Same thing, but creates and retains the new variable:
namcs2019 = var_cross(namcs2019, "Age x Sex", "AGER", "SEX")
tab(namcs2019, "Age x Sex")

# What are the levels or MAJOR?
tab(namcs2019, "MAJOR")
# Tabulate AGER by only 2 of the levels of MAJOR
tab_subset(namcs2019, "AGER", "MAJOR"
, lvls = c("Chronic problem, routine", "Chronic problem, flare-up"))
}
\seealso{
Other tables: 
\code{\link{tab}()},
\code{\link{total}()}
}
\concept{tables}
