% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_rate.R
\name{total_rate}
\alias{total_rate}
\title{Overall rate}
\usage{
total_rate(pop, per = getOption("surveytable.rate_per"))
}
\arguments{
\item{pop}{population}

\item{per}{calculate rate per this many items in the population}
}
\value{
A table
}
\description{
Overall rate
}
\examples{
set_survey(namcs2019sv)
total_rate(uspop_example$total)
}
\seealso{
Other tables: 
\code{\link{tab}()},
\code{\link{tab_cross}()},
\code{\link{tab_rate}()},
\code{\link{tab_subset_rate}()},
\code{\link{total}()}
}
\concept{tables}
