% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{uspop2019}
\alias{uspop2019}
\title{US Population in 2019}
\format{
An object of class \code{list} of length 7.
}
\usage{
uspop2019
}
\description{
Population estimates of the civilian non-institutional population of the
United States as of July 1, 2019. Used for calculating rates. For usage
examples, see the \verb{*_rate} functions.
}
\keyword{datasets}
