% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_all.R
\name{var_all}
\alias{var_all}
\title{Are all the variables true? (Logical AND)}
\usage{
var_all(newvr, vrs)
}
\arguments{
\item{newvr}{name of the new variable to be created}

\item{vrs}{vector of logical variables}
}
\value{
Survey object
}
\description{
Create a new variable which is true if all of the variables in a list of
variables are true.
}
\examples{
set_survey(namcs2019sv)
var_all("Medicare and Medicaid", c("PAYMCARE", "PAYMCAID"))
tab("Medicare and Medicaid")
}
\seealso{
Other variables: 
\code{\link{var_any}()},
\code{\link{var_case}()},
\code{\link{var_collapse}()},
\code{\link{var_copy}()},
\code{\link{var_cross}()},
\code{\link{var_cut}()},
\code{\link{var_not}()}
}
\concept{variables}
