% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_any.R
\name{var_any}
\alias{var_any}
\title{Is any variable true?}
\usage{
var_any(newvr, vrs)
}
\arguments{
\item{newvr}{name of the new variable to be created}

\item{vrs}{vector of logical variables}
}
\value{
(Nothing.)
}
\description{
Create a new variable which is true if any of the variables in a list of
variables are true.
}
\examples{
set_survey("vars2019")
var_any("Imaging services"
, c("ANYIMAGE", "BONEDENS", "CATSCAN", "ECHOCARD", "OTHULTRA"
, "MAMMO", "MRI", "XRAY", "OTHIMAGE"))
tab("Imaging services")
}
\seealso{
Other variables: 
\code{\link{var_collapse}()},
\code{\link{var_copy}()},
\code{\link{var_cross}()},
\code{\link{var_cut}()}
}
\concept{variables}
