% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_not.R
\name{var_not}
\alias{var_not}
\title{Logical NOT}
\usage{
var_not(newvr, vr)
}
\arguments{
\item{newvr}{name of the new variable to be created}

\item{vr}{a logical variable}
}
\value{
Survey object
}
\description{
Logical NOT
}
\examples{
set_survey(namcs2019sv)
var_not("Private insurance not used", "PAYPRIV")
}
\seealso{
Other variables: 
\code{\link{var_all}()},
\code{\link{var_any}()},
\code{\link{var_case}()},
\code{\link{var_collapse}()},
\code{\link{var_copy}()},
\code{\link{var_cross}()},
\code{\link{var_cut}()}
}
\concept{variables}
