% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_list.R
\name{var_num}
\alias{var_num}
\title{Number of variables in a survey design.}
\usage{
var_num(design, out = opts$out$fname)
}
\arguments{
\item{design}{survey design}

\item{out}{(optional) file name for output}
}
\value{
\code{huxtable}
}
\description{
Number of variables in a survey design.
}
\examples{
var_num(namcs2019)
}
