% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{vars2019}
\alias{vars2019}
\title{Selected variables from the National Ambulatory Medical Care Survey (NAMCS) 2019 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 8250 rows and 36 columns.
}
\source{
\itemize{
\item SAS data: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/sas/namcs2019_sas.zip}
\item Survey design variables: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/sas/readme2019-sas.txt}
\item SAS formats: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/sas/nam19for.txt}
\item Documentation: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/doc2019-508.pdf}
\item National Summary Tables: \url{https://www.cdc.gov/nchs/data/ahcd/namcs_summary/2019-namcs-web-tables-508.pdf}
}
}
\usage{
vars2019
}
\description{
Selected variables from a data system of visits to office-based physicians. Note that the unit of
observation is visits, not patients - this distinction is important since a single
patient can make multiple visits.
}
\details{
A survey design object with 8,250 observations.
}
\keyword{datasets}
