% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_import.R
\name{import_sas}
\alias{import_sas}
\title{Import SAS data into R}
\usage{
import_sas(fname_data, fname_fd)
}
\arguments{
\item{fname_data}{SAS survey data file}

\item{fname_fd}{SAS formats data file (produced with CNTLOUT option of PROC FORMAT)}
}
\value{
\code{data.frame}. If it's a complex survey, use \code{svydesign} to create a survey design object.
}
\description{
Import SAS data into R
}
\examples{
\dontrun{
d1 = import_sas(fname_data, fname_fd)
}
}
\seealso{
Other import: 
\code{\link{import_sas_namcs2019puf}()}
}
\concept{import}
