% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_import.R
\name{import_sas_namcs2019puf}
\alias{import_sas_namcs2019puf}
\title{Import SAS data into R, with weighting variables the same as those in NAMCS 2019 PUF.}
\usage{
import_sas_namcs2019puf(fname_data, fname_fd, fname_out)
}
\arguments{
\item{fname_data}{SAS survey data file}

\item{fname_fd}{SAS formats data file (produced with \code{CNTLOUT} option of \verb{PROC FORMAT})}

\item{fname_out}{name of a new R data file}
}
\value{
(Nothing.)
}
\description{
Import SAS data into R, with weighting variables the same as those in NAMCS 2019 PUF.
}
\examples{
\dontrun{
import_sas_namcs2019puf(
fname_data = "../data/namcs2019_sas.sas7bdat"
, fname_fd = "../data/formats_dataset.sas7bdat"
, fname_out = "../data/namcs_2019_puf.rds")
}
}
\seealso{
Other import: 
\code{\link{.import_sas}()}
}
\concept{import}
