% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{namcs2019sv}
\alias{namcs2019sv}
\alias{namcs2019sv_df}
\title{Selected variables from the National Ambulatory Medical Care Survey (NAMCS) 2019 Public Use File (PUF)}
\format{
An object of class \code{survey.design2} (inherits from \code{survey.design}) with 8250 rows and 33 columns.

An object of class \code{data.frame} with 8250 rows and 33 columns.
}
\source{
\itemize{
\item SAS data: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/sas/namcs2019_sas.zip}
\item Survey design variables: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/sas/readme2019-sas.txt}
\item SAS formats: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/sas/nam19for.txt}
\item Documentation: \url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/NAMCS/doc2019-508.pdf}
\item National Summary Tables: \url{https://www.cdc.gov/nchs/data/ahcd/namcs_summary/2019-namcs-web-tables-508.pdf}
}
}
\usage{
namcs2019sv

namcs2019sv_df
}
\description{
Selected variables from a data system of visits to office-based physicians.
Note that the unit of observation is visits, not patients - this distinction
is important since a single patient can make multiple visits.
}
\details{
\code{namcs2019sv_df} is a data frame.

\code{namcs2019sv} is a survey object created from \code{namcs2019sv_df}
using \verb{[survey::svydesign()]}.
}
\keyword{datasets}
