% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.surveytable_table.R
\name{print.surveytable_table}
\alias{print.surveytable_table}
\alias{print.surveytable_list}
\title{Print surveytable tables}
\usage{
\method{print}{surveytable_table}(x, .output = NULL, ...)

\method{print}{surveytable_list}(x, ...)
}
\arguments{
\item{x}{an object of class \code{surveytable_table} or \code{surveytable_list}.}

\item{.output}{output type. \code{NULL} = auto-detect.}

\item{...}{ignored}
}
\value{
\code{x} invisibly.
}
\description{
Print surveytable tables
}
\examples{
set_survey(namcs2019sv)
table1 = tab("AGER")
print(table1)
table_many = tab("MDDO", "SPECCAT", "MSA")
print(table_many)
}
