% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_count.R
\name{set_count_1k}
\alias{set_count_1k}
\alias{set_count_int}
\title{Rounding counts}
\usage{
set_count_1k()

set_count_int()
}
\value{
(Nothing.)
}
\description{
Determines how counts should be rounded.
}
\details{
\itemize{
\item \code{set_count_1k()}: round counts to the nearest 1,000.
\item \code{set_count_int()}: round counts to the nearest integer.
}
}
\examples{
set_survey("namcs2019sv")
set_count_int()
total()

set_count_1k()
total()
}
\seealso{
Other options: 
\code{\link{set_output}()},
\code{\link{set_survey}()},
\code{\link{show_options}()},
\code{\link{surveytable-options}}
}
\concept{options}
