% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_survey.R
\name{set_survey}
\alias{set_survey}
\title{Specify the survey from which you would like to tabulate the estimates from}
\usage{
set_survey(survey_name = "")
}
\arguments{
\item{survey_name}{the name of a \code{survey.design} object (in quotation marks)}
}
\value{
(Nothing.)
}
\description{
You need to do this before the other functions, such as \code{\link[=tab]{tab()}}, will work.
}
\examples{
set_survey("vars2019")
}
