% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_survey.R, R/set_mode.R
\name{set_survey}
\alias{set_survey}
\alias{set_mode}
\title{Specify the survey to analyze}
\usage{
set_survey(design, mode = "default", csv = getOption("surveytable.csv"))

set_mode(mode = "default")
}
\arguments{
\item{design}{either a survey object (created with \code{\link[survey:svydesign]{survey::svydesign()}} or
\code{\link[survey:svrepdesign]{survey::svrepdesign()}}); or, for an unweighted survey, a \code{data.frame}.}

\item{mode}{set certain options. See below.}

\item{csv}{name of a CSV file}
}
\value{
\code{set_survey}: info about the survey. \code{set_mode}: nothing.
}
\description{
You must specify a survey before the other functions, such as \code{\link[=tab]{tab()}},
will work. To convert a \code{data.frame} to a survey object, see \code{\link[survey:svydesign]{survey::svydesign()}}
or \code{\link[survey:svrepdesign]{survey::svrepdesign()}}.
}
\details{
Optionally, the survey can have an attribute called \code{label}, which is the
long name of the survey. Optionally, each variable in the survey can have an
attribute called \code{label}, which is the variable's long name.

If you are not sure what the \code{mode} should be, leave it as \code{"default"}. Here is
what \code{mode} does:
\itemize{
\item \code{"general"} or \code{"default"}:
\itemize{
\item Round counts to the nearest integer -- see \code{\link[=set_count_int]{set_count_int()}}.
\item Do not look for low-precision estimates.
\item Percentage CI's: use standard Korn-Graubard CI's.
}
\item \code{"nchs"}:
\itemize{
\item Round counts to the nearest 1,000 -- see \code{\link[=set_count_1k]{set_count_1k()}}.
\item Identify low-precision estimates.
\item Percentage CI's: adjust Korn-Graubard CI's for the number of degrees of freedom, matching the SUDAAN calculation.
}
}
}
\examples{
set_survey(namcs2019sv)
set_mode("general")
}
\seealso{
Other options: 
\code{\link{set_count_1k}()},
\code{\link{set_output}()},
\code{\link{show_options}()},
\code{\link{surveytable-options}}
}
\concept{options}
