% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_survey.R
\name{set_survey}
\alias{set_survey}
\title{Specify the survey to analyze}
\usage{
set_survey(design)
}
\arguments{
\item{design}{either a survey object (\code{survey.design} or \code{svyrep.design}) or a
\code{data.frame} for an unweighted survey.}
}
\value{
Info about the survey.
}
\description{
You need to specify a survey before the other functions, such as \code{\link[=tab]{tab()}},
will work.
}
\details{
Optionally, the survey can have an attribute called \code{label}, which is the
long name of the survey.

Optionally, each variable in the survey can have an attribute called \code{label},
which is the variable's long name.
}
\examples{
set_survey(namcs2019sv)
}
