% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytable.R
\docType{package}
\name{surveytable-package}
\alias{surveytable}
\alias{surveytable-package}
\title{surveytable: Formatted Survey Estimates}
\description{
Formatted Survey Estimates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/surveytable/}
  \item \url{https://github.com/CDCgov/surveytable}
}

}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov}

}
\keyword{internal}
