% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytable.R
\docType{package}
\name{surveytable-package}
\alias{surveytable}
\alias{surveytable-package}
\title{surveytable: Formatted Survey Estimates}
\description{
Short and understandable commands that generate tabulated, formatted, and rounded survey estimates. Mostly a wrapper for the 'survey' package (Lumley (2004) \doi{10.18637/jss.v009.i08} \url{https://CRAN.R-project.org/package=survey}) that identifies low-precision estimates using the National Center for Health Statistics (NCHS) presentation standards (Parker et al. (2017) \url{https://www.cdc.gov/nchs/data/series/sr_02/sr02_175.pdf}, Parker et al. (2023) \doi{10.15620/cdc:124368}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/surveytable/}
  \item \url{https://github.com/CDCgov/surveytable}
}

}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov} (\href{https://orcid.org/0000-0002-6408-7745}{ORCID})

}
\keyword{internal}
