% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytable.R
\docType{package}
\name{surveytable-package}
\alias{surveytable}
\alias{surveytable-package}
\title{surveytable: Streamlined output from the `survey` package}
\description{
Streamlined output from the `survey` package.
}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov}

}
\keyword{internal}
