% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Tabulate variables}
\usage{
tab(
  ...,
  max_levels = getOption("prettysurvey.out.max_levels"),
  screen = getOption("prettysurvey.out.screen"),
  csv = getOption("prettysurvey.out.csv")
)
}
\arguments{
\item{...}{names of variables (in quotes)}

\item{max_levels}{a categorical variable can have at most this many levels. Used to avoid printing huge tables.}

\item{screen}{print to the screen?}

\item{csv}{name of a CSV file}
}
\value{
A list of \code{data.frame} tables or a single \code{data.frame} table.
}
\description{
Operates on categorical and logical variables, and presents the
estimated counts, their standard errors (SEs) and confidence
intervals (CIs), percentages, and their SEs and CIs. Checks
the presentation guidelines for counts and percentages and flags
estimates if, according to the guidelines,
they should be suppressed, footnoted, or reviewed by an analyst.
}
\details{
CIs are calculated at the 95\% confidence level. CIs for
count estimates are the log Student’s t CIs, with adaptations
for complex surveys. CIs for percentage estimates are
the Korn and Graubard CIs.
}
\examples{
set_survey("vars2019")
tab("AGER")
tab("MDDO", "SPECCAT", "MSA")

# Integrate the output into other programming tasks
set_count_int()
df1 = tab("AGER", screen = FALSE)
df1 = within(df1, {RSE = SE / Number})
}
\seealso{
Other tables: 
\code{\link{tab_cross}()},
\code{\link{total}()}
}
\concept{tables}
