% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{tab}
\alias{tab}
\title{Tabulate variables}
\usage{
tab(
  design,
  ...,
  max.levels = getOption("prettysurvey.out.max_levels"),
  screen = getOption("prettysurvey.out.screen"),
  out = getOption("prettysurvey.out.fname")
)
}
\arguments{
\item{design}{survey design}

\item{...}{names of variables (in quotes)}

\item{max.levels}{a categorical variable can have at most this many levels. Used to avoid printing huge tables.}

\item{screen}{print to the screen?}

\item{out}{file name of CSV file}
}
\value{
a list of \code{data.frame} tables.
}
\description{
Operates on categorical and logical variables, and presents the estimated
counts, their standard errors (SEs) and confidence intervals (CIs),
percentages, and their SEs and CIs. Checks the presentation guidelines for
counts and percentages and flags estimates if, according to the guidelines,
they should be suppressed, footnoted, or reviewed by an analyst. CIs are
calculated at the 95\% confidence level. CIs for the percentage estimates are
calculated using the Korn and Graubard method.
}
\examples{
tab(namcs2019, "MDDO", "SPECCAT", "MSA")
}
