% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cross.R, R/tab_subset.R
\name{tab_cross}
\alias{tab_cross}
\alias{var_cross}
\alias{tab_subset}
\title{Tabulate subsets or interactions}
\usage{
tab_cross(
  vr,
  vrby,
  max_levels = getOption("prettysurvey.max_levels"),
  screen = getOption("prettysurvey.screen"),
  csv = getOption("prettysurvey.csv")
)

var_cross(newvr, vr, vrby)

tab_subset(
  vr,
  vrby,
  lvls = c(),
  max_levels = getOption("prettysurvey.max_levels"),
  screen = getOption("prettysurvey.screen"),
  csv = getOption("prettysurvey.csv")
)
}
\arguments{
\item{vr}{variable to tabulate}

\item{vrby}{use this variable to subset the survey}

\item{max_levels}{a categorical variable can have at most this many levels. Used to avoid printing huge tables.}

\item{screen}{print to the screen?}

\item{csv}{name of a CSV file}

\item{newvr}{name of the new variable to be created}

\item{lvls}{(optional) only show these levels of \code{vrby}}
}
\value{
\itemize{
\item \code{tab_subset}: A list of \code{data.frame} tables or a single \code{data.frame} table.
\item \code{tab_cross}: A \code{data.frame} table.
\item \code{var_cross}: (Nothing.)
}
}
\description{
Create subsets of the survey using one variable, and tabulate another
variable within each of the subsets. Interact two variables and tabulate.
}
\details{
\code{tab_subset} creates subsets using the levels of \code{vrby}, and tabulates
\code{vr} in each subset. Optionally, only use the \code{lvls} levels of \code{vrby}.

\code{tab_cross} crosses or interacts \code{vr} and \code{vrby} and tabulates the new
variable. Tables created using \code{tab_subset} and \code{tab_cross} have the same
counts but different percentages. With \code{tab_subset}, percentages within each
subset add up to 100\%. With \code{tab_cross}, percentages across the entire
population add up to 100\%.

\code{var_cross} creates the new variable and updates the survey,
but does not tabulate the new variable. Use \code{\link[=tab]{tab()}} to tabulate it.
}
\examples{
set_survey("vars2019")

# For each SEX, tabulate AGER
tab_subset("AGER", "SEX")

# Same counts as tab_subset(), but different percentages.
tab_cross("AGER", "SEX")

# Same thing, but creates and retains the new variable:
var_cross("Age x Sex", "AGER", "SEX")
tab("Age x Sex")

# What are the levels or MAJOR?
tab("MAJOR")
# Tabulate AGER by only 2 of the levels of MAJOR
tab_subset("AGER", "MAJOR"
, lvls = c("Chronic problem, routine", "Chronic problem, flare-up"))
}
\seealso{
Other tables: 
\code{\link{tab}()},
\code{\link{total}()}
}
\concept{tables}
