% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table2.R
\name{total}
\alias{total}
\title{Total count}
\usage{
total(design, out = opts$out$fname)
}
\arguments{
\item{design}{survey design}

\item{out}{(optional) file name for output}
}
\value{
\code{huxtable}
}
\description{
Total count
}
\examples{
total(namcs2019)
}
