% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\title{Total count}
\usage{
total(
  screen = getOption("prettysurvey.screen"),
  csv = getOption("prettysurvey.csv")
)
}
\arguments{
\item{screen}{print to the screen?}

\item{csv}{name of a CSV file}
}
\value{
\code{data.frame}
}
\description{
Total count
}
\examples{
set_survey("vars2019")
total()
}
\seealso{
Other tables: 
\code{\link{tab_cross}()},
\code{\link{tab}()}
}
\concept{tables}
