% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table2.R
\name{total}
\alias{total}
\title{Total count}
\usage{
total(
  design,
  raw = opts$tab$raw,
  screen = opts$out$screen,
  prefix = opts$out$prefix
)
}
\arguments{
\item{design}{survey design}

\item{raw}{also output raw counts? (Useful for performing further calculations
, such as calculating rates.)}

\item{screen}{print to the screen?}

\item{prefix}{prefix of a file name to send output to}
}
\value{
\code{huxtable}
}
\description{
Total count
}
\examples{
total(namcs2019)
}
