% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\title{Total count}
\usage{
total(csv = getOption("surveytable.csv"))
}
\arguments{
\item{csv}{name of a CSV file}
}
\value{
\code{data.frame}
}
\description{
Total count
}
\examples{
set_survey(namcs2019sv)
total()
}
\seealso{
Other tables: 
\code{\link{tab_cross}()},
\code{\link{tab_rate}()},
\code{\link{tab_subset_rate}()},
\code{\link{tab}()},
\code{\link{total_rate}()}
}
\concept{tables}
