% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table2.R
\name{total}
\alias{total}
\title{Total count}
\usage{
total(
  design,
  screen = getOption("prettysurvey.out.screen"),
  out = getOption("prettysurvey.out.fname")
)
}
\arguments{
\item{design}{survey design}

\item{screen}{print to the screen?}

\item{out}{file name of CSV file}
}
\value{
\code{data.frame}
}
\description{
Total count
}
\examples{
total(namcs2019)
}
\seealso{
Other tables: 
\code{\link{tab_subset}()},
\code{\link{tab}()}
}
\concept{tables}
