% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_util.R
\name{var_cross}
\alias{var_cross}
\title{Cross two variables in a survey design}
\usage{
var_cross(design, newvar, var1, var2)
}
\arguments{
\item{design}{survey design}

\item{newvar}{name of the new variable to be created}

\item{var1}{first variable}

\item{var2}{second variable}
}
\value{
Updated \code{survey.design}
}
\description{
Cross two variables in a survey design
}
\examples{
namcs2019 = var_cross(namcs2019, "Age x Sex", "AGER", "SEX")
tab(namcs2019, "Age x Sex")
}
