% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_list.R
\name{var_list}
\alias{var_list}
\title{List variables in a survey design}
\usage{
var_list(design, sw = "", all = FALSE, out = opts$out$fname)
}
\arguments{
\item{design}{survey design}

\item{sw}{starting characters in variable name (case insensitive)}

\item{all}{print all variables?}

\item{out}{(optional) file name for output}
}
\value{
\code{huxtable}
}
\description{
List variables in a survey design
}
\examples{
var_list(namcs2019, "age")
}
