% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_num.R
\name{var_num}
\alias{var_num}
\title{Number of variables in a survey.}
\usage{
var_num(
  screen = getOption("prettysurvey.out.screen"),
  out = getOption("prettysurvey.out.fname")
)
}
\arguments{
\item{screen}{print to the screen?}

\item{out}{file name of CSV file}
}
\value{
\code{data.frame}
}
\description{
Number of variables in a survey.
}
\examples{
set_survey("vars2019")
var_num()
}
