% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_list.R
\name{var_num}
\alias{var_num}
\title{Number of variables in a survey.}
\usage{
var_num(
  design,
  screen = getOption("prettysurvey.out.screen"),
  out = getOption("prettysurvey.out.fname")
)
}
\arguments{
\item{design}{survey design}

\item{screen}{print to the screen?}

\item{out}{file name of CSV file}
}
\value{
\code{data.frame}
}
\description{
Number of variables in a survey.
}
\examples{
var_num(namcs2019)
}
