% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_list.R
\name{var_num}
\alias{var_num}
\title{Number of variables in a survey design.}
\usage{
var_num(design, screen = opts$out$screen, prefix = opts$out$prefix)
}
\arguments{
\item{design}{survey design}

\item{screen}{print to the screen?}

\item{prefix}{prefix of a file name to send output to}
}
\value{
\code{huxtable}
}
\description{
Number of variables in a survey design.
}
\examples{
var_num(namcs2019)
}
